refactor: move Python modules to .ci directory and improve CI/CD setup
