+++ .ci/listpkgs_aggregator/__init__.py
@@ -1,0 +1,1 @@
+"""Package aggregator module for NurOS packages."""
