+++ .ci/pyproject.toml
@@ -1,0 +1,30 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "listpkgs-aggregator"
+version = "1.0.0"
+description = "Aggregates package metadata from NurOS-Packages organization"
+authors = [
+    {name = "NurOS Team"}
+]
+dependencies = [
+    "requests>=2.28.0,<3.0.0"
+]
+requires-python = ">=3.9"
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0",
+    "black>=22.0",
+    "flake8>=4.0",
+    "mypy>=0.950"
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["listpkgs_aggregator*"]
+
+[project.scripts]
+listpkgs-aggregate = "listpkgs_aggregator.aggregator:main"
