+++ .ci/test_installation.py
@@ -1,0 +1,38 @@
+#!/usr/bin/env python3
+"""Test script to verify the package installation and functionality."""
+
+import subprocess
+import sys
+import os
+
+def test_installation():
+    """Test that the package installs correctly and the command works."""
+    print("Testing package installation...")
+    
+    # Install the package in development mode
+    result = subprocess.run([sys.executable, "-m", "pip", "install", "-e", ".ci/"], 
+                          cwd=os.getcwd(), capture_output=True, text=True)
+    
+    if result.returncode != 0:
+        print(f"Installation failed: {result.stderr}")
+        return False
+    
+    print("Installation successful!")
+    
+    # Test that the command is available
+    result = subprocess.run(["listpkgs-aggregate", "--help"], 
+                          cwd=os.getcwd(), capture_output=True, text=True)
+    
+    # Note: aggregator.py doesn't have --help option, so we expect this to fail
+    # But we're checking that the command exists
+    print("Command is accessible")
+    
+    return True
+
+if __name__ == "__main__":
+    success = test_installation()
+    if success:
+        print("All tests passed!")
+    else:
+        print("Tests failed!")
+        sys.exit(1)
