+++ .nojekyll
@@ -1,0 +1,1 @@
+# Empty file to disable Jekyll processing on GitHub Pages
